+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Always use the internal ExeFS file object when reading it for FullDecrypted
 * Add docstrings to `NCCHFlags`
 * Add dependency on `Pillow>=8.2`
+* Load SMDH icons using Pillow/PIL, stored in new attributes in the `SMDH` class: `icon_small`, `icon_large`
 
 ## v0.4.7 - April 20, 2021
 * Use absolute paths in `CDNReader`